Add archive summary data
Adding code to pull data from https://kenpom.com/archive.php. The user can specify a historical date to pull data for, which reflects what the Kenpom ratings were on a specific day. This is useful for evaluating how ratings change over time. The date needs to be specified in YYYY-MM-DD format.